fix(proxy): a 2xx the gateway served is metered, whatever the number
attemptCandidate routed on `!= 200`, so a chat completion answered 201 or
202 — what an OpenAI-compatible relay or aggregator can return — went to
forwardUpstreamError. That wrote the row as state="failed", never called
recordTokenUsage, and then forwarded the provider's body to the client
anyway. The caller kept a complete answer that charged no tokens against
the virtual key and debited no TPM bucket: a quota bypass reachable by
choosing such a provider.

The tell was proxy.go's own success gate, `>= 200 && < 300`, a range that
could only ever see 200 because everything else had already been routed
away. Two definitions of "a success the gateway served", and the routing
one won.

The pass-through families already route on the range (multimodal.go), so
the fix is to make chat agree rather than to invent a rule: any 2xx now
reaches the normal handler, where that range check becomes live and
meters exactly as a 200 does. A non-200 success keeps its status on the
way out; rewriting 201 to 200 would be this gateway overruling a provider
about its own answer.

Two consequences worth stating rather than burying:

- 204/205 are forwarded as sent, with no invented body. That was the
  reason the old code forwarded every 2xx, and it now hangs off the
  statuses HTTP actually forbids a body on. Keyed on the status, not on
  the body being empty — a 200 with an empty body is a provider that
  answered nothing and must keep failing, which the existing
  TestHandleNonStreamingResponse_EmptyBody caught when the first cut got
  this wrong.
- A 2xx carrying something that is not a completion is now an error
  instead of being forwarded verbatim. The caller asked for a chat
  completion and an OpenAI client cannot parse `accepted` either;
  forwarding an unreadable success is also how the unmetered path stayed
  invisible.

forwardUpstreamError loses its 2xx branch, now unreachable from both call
sites, and with it the one unbounded io.ReadAll in the function: it
existed because truncating a success corrupts it, which is the reason a
success does not belong there at all. Every remaining branch is capped.

Found by the glm52 Strix run on 29edaec0. proxy.go passed the 800-line
ceiling, so the non-streaming half moved to nonstreaming_response.go
unchanged.

Three mutations killed: restoring `!= 200`, dropping the status
preservation, and widening bodilessSuccessStatus to every 2xx.
